changed ajax requests structure to be handled by controllers

+++ src/main/java/com/hameed/springboot/pharmacyms/controller/MedicationsController.java
@@ -5,13 +5,18 @@ import com.hameed.springboot.pharmacyms.service.CategoryService;
 import com.hameed.springboot.pharmacyms.service.MedicationService;
 import com.hameed.springboot.pharmacyms.service.UnitOfMeasureService;
 import com.hameed.springboot.pharmacyms.service.UserService;
-import jakarta.servlet.http.HttpSession;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.StringTrimmerEditor;
+import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.web.firewall.RequestRejectedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Date;
 
@@ -46,41 +51,44 @@ public class MedicationsController {
         dataBinder.registerCustomEditor(String.class, stringTrimmerEditor);
     }
 
-    @GetMapping("/")
-    public String showAllMedications(Model model) {
+
+
+    @GetMapping
+    public String showAllMedications(Model model, @RequestHeader(value = "X-Requested-With", required = false) String requestedWith) {
         model.addAttribute("medication", new Medication());
         model.addAttribute("medications", medicationService.getAllMedications());
         model.addAttribute("categories", categoryService.getAllCategories());
         model.addAttribute("unitOfMeasures", unitOfMeasureService.getAllUnitsOfMeasures());
         model.addAttribute("fragment", "/fragments/medications-frag");
-        return "layout";
-    }
 
-    @GetMapping("/{id}")
-    public String populateMedication(@PathVariable long id, Model model) {
-        Medication medication = medicationService.getMedicationById(id);
-        model.addAttribute("medication", medication);
-        model.addAttribute("categories", categoryService.getAllCategories());
-        model.addAttribute("unitOfMeasures", unitOfMeasureService.getAllUnitsOfMeasures());
-        return "layout";
+        return "XMLHttpRequest".equals(requestedWith) ? "/fragments/medications-frag" : "layout";
     }
 
     @PostMapping("/add")
-    public String addMedication(@ModelAttribute("medication") Medication medication) {
+    public String addMedication(@ModelAttribute("medication") Medication medication, @RequestHeader(value = "X-Requested-With", required = true) String requestedWith) {
+        if (!"XMLHttpRequest".equals(requestedWith)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Resource not found");
+        }
         medicationService.createMedication(medication);
-        return "redirect:/medications/";
+        return "/fragments/medications-frag";
     }
 
     @PostMapping("/update")
-    public String updateMedication(@ModelAttribute("medication") Medication medication) {
+    public String updateMedication(@ModelAttribute("medication") Medication medication, @RequestHeader(value = "X-Requested-With", required = true) String requestedWith) {
+        if (!"XMLHttpRequest".equals(requestedWith)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Resource not found");
+        }
         medicationService.updateMedication(medication);
-        return "redirect:/medications";
+        return "/fragments/medications-frag";
     }
 
-    @DeleteMapping("/{id}")
-    public String deleteMedication(@PathVariable Long id) {
+    @PostMapping("/{id}")
+    public String deleteMedication(@PathVariable Long id, @RequestHeader(value = "X-Requested-With", required = true) String requestedWith) {
+        if (!"XMLHttpRequest".equals(requestedWith)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Resource not found");
+        }
         medicationService.deleteMedication(id);
-        return "redirect:/medications";
+        return "/fragments/medications-frag";
     }
 
 }
